Refactor code to support guided mode and GPT-4
This commit refactors the code to add support for guided mode and GPT-4. The `--guided` flag allows the user to be prompted for confirmation before reviewing and applying changes for each file. Additionally, the `--gpt4` flag enables the use of GPT-4 instead of the default GPT-3.5 for generating the commit message.